requirements.txt
# Requirements cho Trading Bot với Binance + ChatGPT API

# Core libraries
python-binance>=1.0.16
openai>=1.3.5

# Data analysis
pandas>=1.5.0
numpy>=1.20.0
pandas-ta>=0.3.14b0

# Database (built-in với Python)
# sqlite3 - không cần cài đặt

# Logging & Reporting
matplotlib>=3.5.0
plotly>=5.0.0
dash>=2.0.0
Pillow>=9.0.0  # Để hiển thị biểu đồ trong GUI

# Utilities
python-dotenv>=0.19.0
requests>=2.28.0
# datetime - built-in với Python

# Optional: Telegram notifications
# python-telegram-bot>=20.0

